# docs/swagger.yaml
swagger: '2.0'
schemes:
  - http
info:
  version: 1.0.0
  title: 'Predator'
  description: |
    Predator is an open-source performance framework. It manages the entire lifecycle of load testing a server, from creating a test file, running scheduled and on-demand tests, and finally viewing the test results. Developers who want to load their system end-to-end or test a specific service will benefit from this framework. Bootstrapped with a user-friendly UI alongside a simple REST API, Predator simplifies the performance testing regime.
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/tests:
    get:
      tags:
        - tests
      description: Get list of all test and there descriptions.
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/tests_response'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/error_response'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/error_response'
    post:
      tags:
        - tests
      description: Create new test
      parameters:
        - name: test
          in: body
          description: The test to add
          required: true
          schema:
            $ref: '#/definitions/test'
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/create_test_response'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/error_response'
        422:
          description: Unprocessable entity.
          schema:
            $ref: '#/definitions/error_response'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/error_response'
...
  /v1/jobs:
    get:
      tags:
        - jobs
      description: Get list of all jobs.
      parameters:
        - in: query
          name: one_time
          type: boolean
          description: Should return the one time tests (no cron expression)
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/job_response"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/error_response"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/error_response"
    post:
      tags:
        - jobs
      description: Create a new job that will trigger test run
      parameters:
        - name: job
          in: body
          description: The job to add
          required: true
          schema:
            $ref: "#/definitions/create_job_response"
      responses:
        201:
          description: Success
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/error_response"
        409:
          description: Conflict
          schema:
            $ref: "#/definitions/error_response"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/error_response"
...
  /v1/config:
    get:
      tags:
        - config
      description: Get list of saved configoration
      responses:
        200:
          description: Success
        500:
          description: Internal server error
    put:
      tags:
        - config
      description: Add conig data
      parameters:
        - name: config
          in: body
          description: The config data to add
          required: true
          schema:
            $ref: "#/definitions/config_body"
      responses:
        200:
          description: Success
        500:
          description: Internal server error
definitions:
...
  test:
    type: object
    discriminator: type
    required:
      - name
      - description
      - type
    properties:
      name:
        type: string
        description: The name of the test.
      description:
        type: string
        description: A description of the test.
      type:
        type: string
        enum: ['dsl', 'custom']
...
  test_id:
    required:
      - id
    properties:
      id:
        type: string
        format: uuid
        description: The id of the test.
  create_test_response:
    allOf:
      - $ref: '#/definitions/test_id'
    required:
      - revision_id
    properties:
      revision_id:
        type: string
        format: uuid
        description: The id of the revision.
  tests_response:
    allOf:
      - $ref: '#/definitions/create_test_response'
      - $ref: '#/definitions/test_id'
    required:
      - name
      - description
      - type
      - raw_data
      - artillery_test
      - updated_at
    properties:
      updated_at:
        type: string
        format: date_time
        description: The time the test was created/updated
      name:
        type: string
        description: The name of the test.
      description:
        type: string
        description: The description of the test
      type:
        type: string
        enum: [dsl, custom]
        description: the type of the test, custom - an artillery json created by the client, dsl - a test with saved request definitions api.
      raw_data:
        type: array
        items:
          $ref: '#/definitions/scenario'
      artillery_test:
        type: object
        description: The artillery json that represents the test created from the raw data.
  error_response:
    required:
      - message
    properties:
      message:
        type: string
  scenario:
    required:
      - scenario_name
      - steps
    properties:
      scenario_name:
        type: string
        description: The name of the scenario
      steps:
        type: array
        minItems: 1
        items:
          type: object
          required:
            - action
          properties:
            action:
              type: string
            properties:
              type: object
              description: Any additional properties in the body request that should be changed.
            wait:
              type: integer
              description: To wait 'x' seconds before starting the request.
          description: An object that represent an action.
        description: An array of the steps in the scenario
      weight:
        type: number
        description: The weight (in percentage) of the scenario in the test. If not specified, will be calculated.
  job:
    required:
      - test_id
      - arrival_rate
      - duration
    properties:
      variables:
        type: object
        description: "Variables that the artillery uses and can be defined by the user. Supported variables: 1) appId - if specifed merchantId must be a variable too."
        example: {appId: 'App-name', merchantId: 'cd44fc7a-7719-4383-9981-427ff1210553'}
      test_id:
        type: string
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
        description: The id of the test we would like to schedule.
      cron_expression:
        type: string
        description: A cron expression describes when the test should be triggered, mandatory if run_immediately is not true
      run_immediately:
        type: boolean
        description: Should the test be triggered immediately, regarding the cron expression
      webhooks:
        type: array
        items:
          type: string
          description: The url of where to send the webhook with the report information
      arrival_rate:
        type: number
        minimum: 1
        description: How many requests per second to create
      duration:
        type: number
        minimum: 1
        description: For how many seconds the test should run
      ramp_to:
        type: number
        minimum: 1
        description: A linear ramp up where the number of new arrivals increases linearly over the duration of the phase.
      max_virtual_users:
        type: number
        minimum: 1
        description: The number of concurrent users that can execute requests. The arrival rate will split between them
      parallelism:
        type: number
        minimum: 1
        description: The number of runners to deploy upon start of the test. The arrival rate, ramp to and the max virtual users will be split by this value
      emails:
        type: array
        items:
          type: string
          description: The url of where to send the the email with the report information
      environment:
        type: string
        description: In which environment the test will run
  config_body:
    additionalProperties: false
    properties:
      grafana_url:
        type: string
        description: .
      external_address:
        type: string
        description: .
      internal_address:
        type: string
        description: .
      docker_name:
        type: string
        description: .
      job_platform:
        type: string
        description: .
      runner_cpu:
        type: number
        description: .
      runner_memory:
        type: number
        description: .
      metrics_plugin_name:
        type: string
        enum: ['influx,prometheus']
        description: .
      default_email_address:
        type: string
        description: .
      default_webhook_url:
        type: string
        description: .
      influx_metrics:
        additionalProperties: false
        type: object
        description:  .
        required:
          - host
          - username
          - password
          - database
        properties:
          host:
            type: string
            description: .
          username:
            type: string
            description : .
          password:
            type: string
            description: .
          database:
            type: string
            description: .
      prometheus_metrics:
        additionalProperties: false
        type: object
        description: .
        required:
          - push_gateway_url
        properties:
          push_gateway_url:
            type: string
            description: .
          buckets_sizes:
            type: string
            description: .
      smtp_server:
        additionalProperties: false
        type: object
        description: .
        required:
          - host
          - port
          - username
          - password
          - timeout
        properties:
          host:
            type: string
            description: .
          port:
            type: number
            description: .
          username:
            type: string
            description: .
          password:
            type: string
            description: .
          timeout:
            type: number
            description: .
...
  job_response:
    allOf:
      - $ref: "#/definitions/job"
      - properties:
          id:
            type: string
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
            description: Uuid that identifies the job
  create_job_response:
    allOf:
      - $ref: "#/definitions/job_response"
      - properties:
          run_id:
            type: string
            description: If the job is ran immediately, will show the run id.
...